feat: move random accounts to test helpers
