+++ wasmbinding/tokenfactory/queries_test.go
@@ -9,13 +9,14 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
+	"github.com/kiichain/kiichain/v2/app/apptesting"
 	"github.com/kiichain/kiichain/v2/wasmbinding/helpers"
 	wasmbinding "github.com/kiichain/kiichain/v2/wasmbinding/tokenfactory"
 )
 
 // TestFullDenom tests the GetFullDenom function of the token factory
 func TestFullDenom(t *testing.T) {
-	actor := helpers.RandomAccountAddress()
+	actor := apptesting.RandomAccountAddress()
 
 	specs := map[string]struct {
 		addr         string
@@ -66,7 +67,7 @@ func TestFullDenom(t *testing.T) {
 
 // TestDenomAdmin tests the GetTokenfactoryDenomAdmin function of the token factory
 func TestDenomAdmin(t *testing.T) {
-	addr := helpers.RandomAccountAddress()
+	addr := apptesting.RandomAccountAddress()
 	app, ctx := helpers.SetupCustomApp(t, addr)
 
 	// set token creation fee to zero to make testing easier
